Sunrise Enabled Config and Middleware
